requirements.txt
absl-py==0.13.0
altgraph==0.17
appnope==0.1.2
astunparse==1.6.3
backcall==0.2.0
bcolz==1.2.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
decorator==5.0.9
dlib==19.22.99
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
idna==3.2
importlib-metadata==4.6.3
imutils==0.5.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
kiwisolver==1.3.1
macholib==1.14
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.1
onnx==1.8.1
onnxruntime==1.8.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openvino==2021.4.0
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
PyInquirer==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.2
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
termcolor==1.1.0
torch==1.9.0
torchvision==0.10.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
